Implement pay method
The pay method allows an account to pay an amount that will  reduce the interest or balance of that account.

The updating of interest in update_balance method is abstracted into its own method.

A separate variable og_amount is created to reference the original amount borrowed by the account.