Overhaul the main program as an interface rather than a simple demo of methods.
The main program will run as a simple command line interface where it prompts the user for information. It will also allow the user to interact with the program by giving control options (such as "enter 1 to open account").

A helper function clear_delay is implemented.

The program prompts the user if they are a new customer and prints a message accordingly.