Start to implement the interface loop.
The interface loop prompts the user to enter a command that controls the flow of the program. The user can now quit the program succesfully.

Create a "database" dictionary to hold the user's accounts.

clear_delay function calls changed to 1 second instead of 1.5

Fixed the logic for determining if the user is a new customer or not.