Implement payment option.
Entering "2" when promted allows the customer to pay their existing accounts. A list of all their accounts will be printed for them to see the specific IDs. They are then prompted to enter the ID of the account they want to pay.

press_enter method has been made more "modular". It now accepts an optional parameter msg which it prints out as it delays execution. The user input is also returned for the use of the caller.