Restructure the prompt so that the user is no longer asked if they are a returning customer or not.
After asking for a code review in [stackexchange](https://codereview.stackexchange.com/questions/162682/a-command-line-application-that-manages-my-lending-accounts/162701#162701). I agreed that asking the customer if they are a returning one is bad design since the program should handle the lookup. If found, the program runs as usual. If the user is not found, they will be asked if they want to register or not. The program exits if the user declines.